Move basestore scan functitons into new files.
